# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AirFlex API",
    "version": "1.0.0",
    "description": "Open-source P2P Airtime & Data Exchange Marketplace backed by the Stellar Network. All protected endpoints require a Bearer JWT obtained from POST /api/v1/auth/verify-otp.",
    "contact": {
      "name": "AirFlex Core Team",
      "url": "https://github.com/arflexx/Airflex"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by POST /api/v1/auth/verify-otp. Pass as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Descriptive error message"
          }
        }
      },
...
      "WalletResponse": {
        "type": "object",
        "required": [
          "publicKey",
          "balance",
          "asset",
          "network"
        ],
        "properties": {
          "publicKey": {
            "type": "string",
            "example": "GABC1234..."
          },
          "balance": {
            "type": "string",
            "description": "XLM balance as a decimal string.",
            "example": "10000.0000000"
          },
          "asset": {
            "type": "string",
            "example": "XLM"
          },
          "network": {
            "type": "string",
            "example": "testnet"
          }
        }
      },
      "Bank": {
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "058"
          },
          "name": {
            "type": "string",
            "example": "Guaranty Trust Bank"
          }
        }
      },
      "BanksResponse": {
        "type": "object",
        "required": [
          "banks"
        ],
        "properties": {
          "banks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Bank"
            }
          }
        }
      },
      "ResolveAccountResponse": {
        "type": "object",
        "required": [
          "account_name"
        ],
        "properties": {
          "account_name": {
            "type": "string",
            "example": "JOHN DOE"
          }
        }
      },
      "WithdrawBody": {
        "type": "object",
        "required": [
          "amount",
          "bank_code",
          "account_number",
          "account_name"
        ],
        "properties": {
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0,
            "example": 1000
          },
          "bank_code": {
            "type": "string",
            "example": "058"
          },
          "account_number": {
            "type": "string",
            "example": "0123456789"
          },
          "account_name": {
            "type": "string",
            "example": "JOHN DOE"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid Bearer JWT.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Missing or invalid Authorization header"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The requested resource does not exist.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Trade offer not found"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Internal server error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/wallet": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "Get wallet balance",
        "description": "Returns the authenticated user's Stellar public key and current XLM balance fetched from Horizon. The secret key is never returned.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Wallet details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "Horizon RPC unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wallet/banks": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "List Nigerian banks",
        "description": "Returns Paystack's list of Nigerian banks for account number resolution.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Bank list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BanksResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "Paystack API unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wallet/resolve-account": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "Resolve bank account name",
        "description": "Resolves a Nigerian bank account number to the account holder name via Paystack.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "account_number",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "0123456789"
            }
          },
          {
            "name": "bank_code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "058"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account name resolved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveAccountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "Paystack API unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wallet/withdraw": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "Withdraw to bank account",
        "description": "Initiates a withdrawal to a Nigerian bank account via Paystack.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Withdrawal initiated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or insufficient balance.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
...
}